requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
dnspython==1.16.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.0
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
jose==1.0.0
passlib==1.7.4
psutil==5.8.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1